% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{default_lambda_clime}
\alias{default_lambda_clime}
\title{Theory-Scale CLIME Tuning Level}
\usage{
default_lambda_clime(p, N, h)
}
\arguments{
\item{p}{Number of fixed-effect covariates.}

\item{N}{Total sample size.}

\item{h}{Smoothing bandwidth.}
}
\value{
Numeric scalar tuning level.
}
\description{
Computes the default CLIME tuning level
\verb{sqrt(log(p)/(N h)) + h^2}.
}
