% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpDMEQR-package.R
\docType{package}
\name{jpDMEQR-package}
\alias{jpDMEQR}
\alias{jpDMEQR-package}
\title{jpDMEQR: Jointly Penalised and Debiased Inference for Mixed-Effects QR}
\description{
Provides screening, penalised estimation, debiasing, and simulation tools
for high-dimensional mixed-effects quantile regression.
}
\author{
\strong{Maintainer}: Jian Hou \email{houjian@ruc.edu.cn}

}
\keyword{internal}
