% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpDMEQR-package.R
\docType{package}
\name{jpDMEQR-package}
\alias{jpDMEQR}
\alias{jpDMEQR-package}
\title{jpDMEQR: Jointly Penalised and Debiased Inference for Mixed-Effects QR}
\description{
Provides screening, penalised estimation, debiasing, and simulation tools
for high-dimensional mixed-effects quantile regression.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Stork343/jpDMEQR}
  \item Report bugs at \url{https://github.com/Stork343/jpDMEQR/issues}
}

}
\author{
\strong{Maintainer}: Jian Hou \email{houjian@ruc.edu.cn}

}
\keyword{internal}
